InvalidPowerOfTwoForCapacity should take an int

+++ src/main/scala/RingBuffer.scala
@@ -9,7 +9,7 @@ trait RingBuffer[T] {
 }
 
 class InvalidPowerOfTwoForCapacity(msg:String) extends RuntimeException(msg) {
-  def this(powerOfTwoForCapacity: Long) = {
+  def this(powerOfTwoForCapacity: Int) = {
     this("Power of two for RingBuffer capacity must be between 1 and 30 (inclusive). It was %d".format(powerOfTwoForCapacity))
   }
 }
